fix-a-warn
git-svn-id: svn://svn.sphinxsearch.com/sphinx/trunk@749 406a0c4d-033a-0410-8de8-e80135713968

+++ src/sphinx.cpp
@@ -3581,8 +3581,8 @@ BYTE * CSphTokenizer_UTF8::GetTokenSyn ()
 
 		for ( ; m_iSynFlushing<0; )
 		{
-			assert ( m_iSynAccBytes<sizeof(m_sSynAcc) );
-			assert ( m_iSynFolded<sizeof(m_sSynFolded) );
+			assert ( m_iSynAccBytes<(int)sizeof(m_sSynAcc) );
+			assert ( m_iSynFolded<(int)sizeof(m_sSynFolded) );
 
 			// get raw and folded codepoints
 			int iCode = -1;
